+++ src/main/java/itstime/shootit/greme/challenge/dto/response/ChallengeRes.java
@@ -1,8 +1,8 @@
-package itstime.shootit.greme.post.dto;
+package itstime.shootit.greme.challenge.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import itstime.shootit.greme.challenge.dto.GetChallengeList;
-import itstime.shootit.greme.challenge.dto.GetChallengeSummaryRes;
+import itstime.shootit.greme.challenge.dto.response.GetChallengeListRes;
+import itstime.shootit.greme.challenge.dto.response.GetChallengeSummaryRes;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,13 +14,13 @@ import java.util.List;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ChallengeDTO {
+public class ChallengeRes {
 
     @Schema(description = "해당 챌린지에 참여하는지")
     private boolean status;
 
     @Schema(description = "챌린지 참여 목록")
-    private List<GetChallengeList> getChallengeLists;
+    private List<GetChallengeListRes> getChallengeLists;
 
     @Schema(description = "해당 챌린지 정보")
     private GetChallengeSummaryRes summaryRes;
