+++ src/main/java/itstime/shootit/greme/challenge/dto/response/GetChallengeTitleRes.java
@@ -1,4 +1,4 @@
-package itstime.shootit.greme.post.dto;
+package itstime.shootit.greme.challenge.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
