+++ src/main/java/itstime/shootit/greme/challenge/presentation/ChallengeController.java
@@ -3,11 +3,11 @@ package itstime.shootit.greme.challenge.presentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import itstime.shootit.greme.challenge.application.ChallengeService;
-import itstime.shootit.greme.challenge.dto.GetChallengeSummaryRes;
+import itstime.shootit.greme.challenge.dto.response.GetChallengeSummaryRes;
 import itstime.shootit.greme.challenge.dto.ChallengeTitle;
 import itstime.shootit.greme.oauth.application.JwtTokenProvider;
-import itstime.shootit.greme.post.dto.ChallengeDTO;
-import itstime.shootit.greme.user.dto.GetProfileRes;
+import itstime.shootit.greme.challenge.dto.response.ChallengeRes;
+import itstime.shootit.greme.user.dto.response.GetProfileRes;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -68,18 +68,11 @@ public class ChallengeController {
         return challengeService.showUserProfile(jwtTokenProvider.getEmail(accessToken), user_id);
     }
 
-//    @Operation(summary = "챌린지 클릭시 챌린지 정보 & 참여 목록 조회하기", parameters = {@Parameter(name = "accessToken", description = "액세스 토큰"),
-//        @Parameter(name = "challengeId", description = "챌린지 고유 id값")})
-//    @GetMapping("/{challengeId}")
-//    public ChallengeDTO showChallengeList(@PathVariable Long challengeId, @RequestHeader("accessToken") String accessToken){
-//        return challengeService.showChallengeList(jwtTokenProvider.getEmail(accessToken), challengeId);
-//
-//    }
-
-    @GetMapping("/{challenge_id}/vv")
-    public ChallengeDTO showChallengeList(@PathVariable Long challenge_id){
-        System.out.println("들어와?");
-        return challengeService.showChallengeList("test@naver.com", challenge_id);
+    @Operation(summary = "챌린지 클릭시 챌린지 정보 & 참여 목록 조회하기", parameters = {@Parameter(name = "accessToken", description = "액세스 토큰"),
+        @Parameter(name = "challengeId", description = "챌린지 고유 id값")})
+    @GetMapping("/{challengeId}")
+    public ChallengeRes showChallengeList(@PathVariable Long challengeId, @RequestHeader("accessToken") String accessToken){
+        return challengeService.showChallengeList(jwtTokenProvider.getEmail(accessToken), challengeId);
 
     }
