+++ src/main/java/itstime/shootit/greme/post/infrastructure/PostRepository.java
@@ -1,20 +1,14 @@
 package itstime.shootit.greme.post.infrastructure;
 
 import itstime.shootit.greme.post.domain.Post;
-import itstime.shootit.greme.post.dto.GetChallengeTitleRes;
-import itstime.shootit.greme.post.dto.GetPostSummaryRes;
-import itstime.shootit.greme.post.dto.GetPostRes;
+import itstime.shootit.greme.post.dto.response.GetPostSummaryRes;
+import itstime.shootit.greme.post.dto.response.GetPostRes;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
-
-    @Query(value = "SELECT c.title FROM ChallengePost cp LEFT OUTER JOIN Challenge c on c.id = cp.challenge_id " +
-            "LEFT OUTER JOIN Post p on p.id = cp.post_id WHERE p.id = :post_id", nativeQuery = true)
-    GetChallengeTitleRes findChallengeTitle(Long post_id);
-
 
     @Query(value = "SELECT p.id, p.image FROM Post p WHERE p.user_id = :userId AND p.status = true ORDER BY p.createdDate DESC LIMIT 5;", nativeQuery = true)
     List<GetPostRes> findRecentPostByUserEmail(Long userId);
