+++ src/main/java/itstime/shootit/greme/post/presentation/PostController.java
@@ -8,7 +8,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import itstime.shootit.greme.aws.application.S3Uploader;
 import itstime.shootit.greme.oauth.application.JwtTokenProvider;
 import itstime.shootit.greme.post.application.PostService;
-import itstime.shootit.greme.post.dto.GetShowPostRes;
+import itstime.shootit.greme.post.dto.response.GetShowPostRes;
 import itstime.shootit.greme.post.dto.request.CreationReq;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
